backend/lucerna/requirements.txt
Django==5.0.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
python-dotenv==1.0.0
djangorestframework==3.14.0
PyJWT==2.8.0
django-cors-headers